VM: Update for new transactionExecutor API

+++ src/main/java/org/semux/vm/client/SemuxBlockStore.java
@@ -22,7 +22,7 @@ public class SemuxBlockStore implements BlockStore {
     }
 
     @Override
-    public byte[] getBlockHashByNumber(int index) {
+    public byte[] getBlockHashByNumber(long index) {
         return blockchain.getBlock(index).getHash();
     }
 }
